app/requirements.txt
Flask
mysql-connector
requests
python-dotenv